+++ PickFLPrj/src/com/pickfl/admin/controller/AdminMemberModify.java
@@ -1,4 +1,4 @@
-package com.kh.inquery.controller;
+package com.pickfl.admin.controller;
 
 import java.io.IOException;
 
@@ -8,11 +8,10 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet("/inquery")
-public class InqueryController extends HttpServlet{
-	
+@WebServlet("/admin_modify")
+public class AdminMemberModify extends HttpServlet{
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-		req.getRequestDispatcher("/WEB-INF/views/inquery/inquery.jsp").forward(req, resp);
+		req.getRequestDispatcher("/WEB-INF/views/admin/memberModify.jsp").forward(req, resp);
 	}
 }
