+++ PickFLPrj/src/com/pickfl/inquery/controller/InqueryWriteController.java
@@ -1,4 +1,4 @@
-package com.kh.admin.controller;
+package com.pickfl.inquery.controller;
 
 import java.io.IOException;
 
@@ -8,10 +8,11 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet("/memberInfo")
-public class MemberInfoController extends HttpServlet {
+@WebServlet("/inqueryWrite")
+public class InqueryWriteController extends HttpServlet{
+	
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-		req.getRequestDispatcher("/WEB-INF/views/admin/memberInfo.jsp").forward(req, resp);
+		req.getRequestDispatcher("/WEB-INF/views/inquery/inqueryWrite.jsp").forward(req, resp);
 	}
 }
